refactor: fold the TLS retry test into runner_test.go
Tests for runner.go belong in runner_test.go; a per-scenario file drifts
away from the code it covers. Also cut the four-line preamble on the
retry down to the one fact that is not already on the next line.

Claude-Local-Session: claude --resume 0fe5b680-5ba5-43e6-b354-43d4388aaf9c
